Deploy a Configurable Web Server
Add an additional output variable 'public_ip'.
This is particularly handy for scripting. For example, you could
create a deployment script that runs `terraform apply' to deploy
the web server, uses `terraform output public_ip' to grab its
public IP, and runs `curl' on the IP as a quick smoke test to
validate that the deployment worked.